+++ CHANGELOG.md
@@ -6,7 +6,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 - Feature: Make `CrowdfundingCard` a function instead of a `PureComponent`.
 - Feature: Add `imageContainerRatio` prop to `CrowdfundingCard` component.
-- Fix: Hide `additionalInfo` when `loading` `CrowdfundingCard` prop is `true`.
+- Fix: On `CrowdfundingCard`, hide `additionalInfo` when `loading` prop is `true`.
 
 ## [2.72.0] - 2020-05-28